Homework 4
Completed the homework assignment by changing to C++ and outputing to a file

+++ Homework4/src/Room.h
@@ -14,13 +14,13 @@
 typedef struct {
 	int roomNum;
 	bool isOpen;
-	int treasure;
+	double treasure;
 } Room;
 
 
 typedef struct {
-	int treasure;
-	int maxTreasure;
+	double treasure;
+	double maxTreasure;
 	int nRoomSearched;
 	int maxRooms;
 	Room *roomsSearched; //define as array of size maxRooms
